Move commit model into its own file
Let's stick to the standard scheme - otherwise we might as well just
have a single models file.
