+++ pydis_site/apps/content/models/__init__.py
@@ -1,3 +1,4 @@
-from .tag import Commit, Tag
+from .commit import Commit
+from .tag import Tag
 
 __all__ = ["Commit", "Tag"]
